backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-jose[cryptography]==3.3.0
cryptography==43.0.0
torch==2.4.0
transformers==4.44.0
accelerate==0.33.0
bitsandbytes==0.43.0
captum==0.7.0
pydantic==2.8.0
pytest==8.3.0
pytest-asyncio==0.23.0
httpx==0.27.0
